- Break out `GraphvizTools` into a separate Mill module, to avoid bloating the main jar and try to shorten the `MILL_SCALA_WORKER` classpath being sent to scalajslib.test (which is blowing up on windows as the CLI command is too long)
- Move the meat of `resolveDependencies` from `scalalib` to `main`, to support resolving mill modules

- DRY up resolution of mill modules in `Util.millProjectModule`
