+++ scalajslib/src/mill/scalajslib/ScalaJSModule.scala
@@ -27,19 +27,10 @@ trait ScalaJSModule extends scalalib.ScalaModule { outer =>
 
   def sjsBridgeClasspath = T {
     val jsBridgeKey = "MILL_SCALAJS_BRIDGE_" + scalaJSBridgeVersion().replace('.', '_')
-    val jsBridgePath = sys.props(jsBridgeKey)
-    if (jsBridgePath != null) Success(
-      Agg(PathRef(Path(jsBridgePath), quick = true))
-    ) else resolveDependencies(
-      Seq(
-        Cache.ivy2Local,
-        MavenRepository("https://repo1.maven.org/maven2"),
-        MavenRepository("https://oss.sonatype.org/content/repositories/releases")
-      ),
-      Lib.depToDependency(_, "2.12.4", ""),
-      Seq(
-        ivy"com.lihaoyi::mill-scalajslib-jsbridges-${scalaJSBridgeVersion()}:${sys.props("MILL_VERSION")}"
-      )
+    mill.modules.Util.millProjectModule(
+      jsBridgeKey,
+      s"mill-scalajslib-jsbridges-${scalaJSBridgeVersion()}",
+      repositories
     ).map(_.filter(_.path.toString.contains("mill-scalajslib-jsbridges")))
   }
